Split backend.rs into logical file structure
In order to make code more readable and make development of
various backends easier, backend.rs is split info multiple
submodules.
